[refactor] 타입 선언 코드 관련 별도로 파읿 분리
