+++ src/components/left-side/OverviewStats/OverviewStats.ts
@@ -1,0 +1,10 @@
+export interface OverviewStats {
+	championName: string;
+	championMinionKillAverage: number;
+	championMinionKillPerMin: number;
+	kill: number;
+	death: number;
+	assist: number;
+	winRatio: number;
+	playTimes: number;
+}
